requirements.txt
absl-py==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
proglog==0.1.10
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sympy==1.12
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
